+++ projects/angular-auth-oidc-client/src/lib/config/openid-configuration.ts
@@ -18,6 +18,13 @@ export interface OpenIdConfiguration {
   authWellknownEndpointUrl?: string;
   authWellknownEndpoints?: AuthWellKnownEndpoints;
 
+    /** 
+     * Override the default Security Token Service wellknown endpoint postfix. 
+     * 
+     * @default /.well-known/openid-configuration
+     */
+  authWellknownUrlSuffix?: string;
+
   /** The redirect URL defined on the Security Token Service. */
   redirectUrl?: string;
   /**