+++ projects/angular-auth-oidc-client/src/lib/interceptor/auth.interceptor.spec.ts
@@ -3,12 +3,9 @@ import {
   HttpClient,
   provideHttpClient,
   withInterceptors,
+  withInterceptorsFromDi,
 } from '@angular/common/http';
-import {
-  HttpClientTestingModule,
-  HttpTestingController,
-  provideHttpClientTesting,
-} from '@angular/common/http/testing';
+import { HttpTestingController, provideHttpClientTesting } from '@angular/common/http/testing';
 import { TestBed, waitForAsync } from '@angular/core/testing';
 import { mockProvider } from '../../test/auto-mock';
 import { AuthStateService } from '../auth-state/auth-state.service';
@@ -27,7 +24,7 @@ describe(`AuthHttpInterceptor`, () => {
   describe(`with Class Interceptor`, () => {
     beforeEach(() => {
       TestBed.configureTestingModule({
-        imports: [HttpClientTestingModule],
+        imports: [],
         providers: [
           ClosestMatchingRouteService,
           {
@@ -38,6 +35,8 @@ describe(`AuthHttpInterceptor`, () => {
           mockProvider(AuthStateService),
           mockProvider(LoggerService),
           mockProvider(ConfigurationService),
+          provideHttpClient(withInterceptorsFromDi()),
+          provideHttpClientTesting(),
         ],
       });
 
@@ -234,7 +233,7 @@ describe(`AuthHttpInterceptor`, () => {
       ]);
       spyOn(
         closestMatchingRouteService,
-        'getConfigIdForClosestMatchingRoute'
+        'getConfigIdForClosestMatchingRoute',
       ).and.returnValue({
         matchingRoute: null,
         matchingConfig: null,