+++ projects/sample-code-flow-standalone/src/app/navigation/navigation.component.ts
@@ -1,5 +1,5 @@
-import { NgIf } from '@angular/common';
-import { Component, OnInit, inject } from '@angular/core';
+
+import { Component, inject } from '@angular/core';
 import { RouterLink } from '@angular/router';
 import { OidcSecurityService } from 'angular-auth-oidc-client';
 
@@ -8,22 +8,11 @@ import { OidcSecurityService } from 'angular-auth-oidc-client';
   templateUrl: 'navigation.component.html',
   styleUrls: ['navigation.component.css'],
   standalone: true,
-  imports: [RouterLink, NgIf],
+  imports: [RouterLink],
 })
-export class NavigationComponent implements OnInit {
+export class NavigationComponent  {
   private readonly oidcSecurityService = inject(OidcSecurityService);
-
-  isAuthenticated = false;
-
-  ngOnInit(): void {
-    this.oidcSecurityService.isAuthenticated$.subscribe(
-      ({ isAuthenticated }) => {
-        this.isAuthenticated = isAuthenticated;
-
-        console.warn('authenticated: ', isAuthenticated);
-      }
-    );
-  }
+  protected readonly authenticated = this.oidcSecurityService.authenticated;
 
   login(): void {
     this.oidcSecurityService.authorize();