+++ projects/schematics/src/ng-add/actions/add-dependencies.ts
@@ -5,7 +5,7 @@ import { NgAddOptions } from '../models/ng-add-options';
 const dependenciesToAdd: any[] = [
   {
     name: 'angular-auth-oidc-client',
-    version: '17.0.0',
+    version: '18.0.1',
   },
 ];