+++ projects/schematics/src/ng-add/actions/copy-standalone-file.ts
@@ -18,7 +18,7 @@ import { AUTH_0, AZURE_AD_REFRESH_TOKENS, AZURE_AD_SILENT_RENEW, DEFAULT_CONFIG,
 
 export function copyStandaloneFile(options: NgAddOptions): Rule {
   return (host: Tree, context: SchematicContext) => {
-    const project = getProject(host);
+    const [,project] = getProject(host);
 
     const { fileName, filesFolder } = options.standaloneInfo!;